Mv utilities from notebooks and split it
